touch the code again
isolate the regions of the array that don't have spots, find average intensity. that is generally the smearing.

rearrange the sorting algorithm so that the left to right, top to bottom sorter is a separate function so it can be used independently.

optimize circle identification algorithm to find all circles every time. make a script that optimizes this, compare to a standard where I count them myself.

compare the mean intensity background smearing algorithm to machine learning model